cloud-functions/iris-webhook/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-aiplatform==1.*
Flask==3.*
